missing casts wanted by clang

+++ libisc/errno2result.c
@@ -98,7 +98,7 @@ isc___errno2result(int posixerrno, const char *file, unsigned int line) {
 		return (ISC_R_CONNREFUSED);
 	default:
 		ISC_IGNORE(strerror_r(posixerrno, strbuf, sizeof(strbuf)));
-		UNEXPECTED_ERROR(file, line, "unable to convert errno "
+		UNEXPECTED_ERROR(file, (int)line, "unable to convert errno "
 				 "to isc_result: %d: %s",
 				 posixerrno, strbuf);
 		/*
